[skip ci] 8214: added submission and validation actions and sequences, removed old modal sequences and actions for editing petitioner counsel
