+++ web-client/src/presenter/sequences/gotoEditPetitionerCounselSequence.js
@@ -2,7 +2,7 @@ import { clearFormAction } from '../actions/clearFormAction';
 import { getCaseAction } from '../actions/getCaseAction';
 import { setCaseAction } from '../actions/setCaseAction';
 import { setCurrentPageAction } from '../actions/setCurrentPageAction';
-import { setPetitionerCounselFormAction } from '../actions/caseDetailEdit/setPetitionerCounselFormAction';
+import { setPetitionerCounselFormAction } from '../actions/caseAssociation/setPetitionerCounselFormAction';
 import { stopShowValidationAction } from '../actions/stopShowValidationAction';
 
 export const gotoEditPetitionerCounselSequence = [
